refactor(farmer): split FarmerTacticalPlan (599→97) em hook + componentes
Extrai verbatim de src/pages/FarmerTacticalPlan.tsx:
- useFarmerTacticalPlan: state, loadCustomers, geração com checagem de
  eficiência e handlers (copy/toggle/filtro)
- GerarPlanoCard, EfficiencyAlertDialog, PlanCard, RecordResultDialog,
  PlanSection (Section/MetricRow/CopyButton), types.ts, config.ts
- +15 testes dos pedaços presentacionais

Comportamento preservado 1:1. tsc baseline=0, eslint=0 erros
(1 warning exhaustive-deps pré-existente, apenas relocado para o hook).
